Moved JobEngine exceptions to API
