+++ job-engine/api/src/main/java/org/eclipse/kapua/job/engine/exception/CheckRunningJobEngineException.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018, 2019 Eurotech and/or its affiliates and others
+ * Copyright (c) 2019 Eurotech and/or its affiliates and others
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -9,21 +9,21 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.job.engine.jbatch.exception;
+package org.eclipse.kapua.job.engine.exception;
 
 import org.eclipse.kapua.KapuaException;
 
-public abstract class JobEngineException extends KapuaException {
+public class CheckRunningJobEngineException extends KapuaException {
 
-    protected JobEngineException(KapuaJobEngineErrorCodes code) {
+    protected CheckRunningJobEngineException(KapuaJobEngineErrorCodes code) {
         super(code);
     }
 
-    protected JobEngineException(KapuaJobEngineErrorCodes code, Object... arguments) {
+    protected CheckRunningJobEngineException(KapuaJobEngineErrorCodes code, Object... arguments) {
         super(code, arguments);
     }
 
-    protected JobEngineException(KapuaJobEngineErrorCodes code, Throwable cause, Object... arguments) {
+    protected CheckRunningJobEngineException(KapuaJobEngineErrorCodes code, Throwable cause, Object... arguments) {
         super(code, cause, arguments);
     }
