+++ job-engine/api/src/main/java/org/eclipse/kapua/job/engine/exception/JobCheckRunningException.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018 Eurotech and/or its affiliates and others
+ * Copyright (c) 2019 Eurotech and/or its affiliates and others
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -9,13 +9,13 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.job.engine.jbatch.exception;
+package org.eclipse.kapua.job.engine.exception;
 
 import org.eclipse.kapua.model.id.KapuaId;
 
-public class JobNotRunningException extends JobEngineException {
+public class JobCheckRunningException extends CheckRunningJobEngineException {
 
-    public JobNotRunningException(KapuaId scopeId, KapuaId jobId) {
-        super(KapuaJobEngineErrorCodes.JOB_NOT_RUNNING, scopeId, jobId);
+    public JobCheckRunningException(Throwable t, KapuaId scopeId, KapuaId jobId) {
+        super(KapuaJobEngineErrorCodes.JOB_CHECK_RUNNING, t, scopeId, jobId);
     }
 }
