+++ job-engine/api/src/main/java/org/eclipse/kapua/job/engine/exception/JobNotRunningException.java
@@ -9,13 +9,13 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.job.engine.jbatch.exception;
+package org.eclipse.kapua.job.engine.exception;
 
 import org.eclipse.kapua.model.id.KapuaId;
 
-public class JobStartingException extends JobEngineException {
+public class JobNotRunningException extends JobEngineException {
 
-    public JobStartingException(Throwable t, KapuaId scopeId, KapuaId jobId) {
-        super(KapuaJobEngineErrorCodes.JOB_STARTING, t, scopeId, jobId);
+    public JobNotRunningException(KapuaId scopeId, KapuaId jobId) {
+        super(KapuaJobEngineErrorCodes.JOB_NOT_RUNNING, scopeId, jobId);
     }
 }
