+++ job-engine/api/src/main/java/org/eclipse/kapua/job/engine/exception/JobResumingException.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018 Eurotech and/or its affiliates and others
+ * Copyright (c) 2019 Eurotech and/or its affiliates and others
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -9,13 +9,17 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.job.engine.jbatch.exception;
+package org.eclipse.kapua.job.engine.exception;
 
 import org.eclipse.kapua.model.id.KapuaId;
 
-public class JobMissingTargetException extends JobEngineException {
+public class JobResumingException extends JobEngineException {
 
-    public JobMissingTargetException(KapuaId scopeId, KapuaId jobId) {
-        super(KapuaJobEngineErrorCodes.JOB_TARGET_MISSING, scopeId, jobId);
+    public JobResumingException(Throwable t, KapuaId scopeId, KapuaId jobId) {
+        this(t, scopeId, jobId, null);
+    }
+
+    public JobResumingException(Throwable t, KapuaId scopeId, KapuaId jobId, KapuaId jobExecutionId) {
+        super(KapuaJobEngineErrorCodes.JOB_RESUMING, t, scopeId, jobId, jobExecutionId);
     }
 }
